'major features added'

+++ src/main/java/com/krakatoaventures/distributionapi/entity/Owner.java
@@ -11,7 +11,7 @@ public class Owner {
     private String name;
     private int commitment;
     private Map<String, Integer> sharesByClass;
-    private int distribution;
+    private double distribution;
 
     public int sumUnits() {
         int totalUnits = sharesByClass.values().stream().mapToInt(i->i).sum();
